Sort all borrowck errors within a single typeck owner together
