+++ tests/ui/async-await/issues/issue-62097.stderr
@@ -1,3 +1,17 @@
+error[E0521]: borrowed data escapes outside of method
+  --> $DIR/issue-62097.rs:13:9
+   |
+LL |     pub async fn run_dummy_fn(&self) {
+   |                               -----
+   |                               |
+   |                               `self` is a reference that is only valid in the method body
+   |                               let's call the lifetime of this reference `'1`
+LL |         foo(|| self.bar()).await;
+   |         ^^^^^^^^^^^^^^^^^^
+   |         |
+   |         `self` escapes the method body here
+   |         argument requires that `'1` must outlive `'static`
+
 error[E0373]: closure may outlive the current function, but it borrows `self`, which is owned by the current function
   --> $DIR/issue-62097.rs:13:13
    |
@@ -15,20 +29,6 @@ help: to force the closure to take ownership of `self` (and any other referenced
    |
 LL |         foo(move || self.bar()).await;
    |             ++++
-
-error[E0521]: borrowed data escapes outside of method
-  --> $DIR/issue-62097.rs:13:9
-   |
-LL |     pub async fn run_dummy_fn(&self) {
-   |                               -----
-   |                               |
-   |                               `self` is a reference that is only valid in the method body
-   |                               let's call the lifetime of this reference `'1`
-LL |         foo(|| self.bar()).await;
-   |         ^^^^^^^^^^^^^^^^^^
-   |         |
-   |         `self` escapes the method body here
-   |         argument requires that `'1` must outlive `'static`
 
 error: aborting due to 2 previous errors
