+++ tests/ui/borrowck/borrowck-multiple-captures.stderr
@@ -78,16 +78,6 @@ help: consider cloning the value if the performance cost is acceptable
 LL |     drop(x2.clone());
    |            ++++++++
 
-error[E0382]: use of moved value: `x`
-  --> $DIR/borrowck-multiple-captures.rs:41:14
-   |
-LL |         drop(x);
-   |              - value moved here
-LL |         drop(x);
-   |              ^ value used here after move
-   |
-   = note: move occurs because `x` has type `Box<i32>`, which does not implement the `Copy` trait
-
 error[E0505]: cannot move out of `x` because it is borrowed
   --> $DIR/borrowck-multiple-captures.rs:38:19
    |
@@ -110,7 +100,7 @@ LL |     let p = &x.clone();
    |               ++++++++
 
 error[E0382]: use of moved value: `x`
-  --> $DIR/borrowck-multiple-captures.rs:52:14
+  --> $DIR/borrowck-multiple-captures.rs:41:14
    |
 LL |         drop(x);
    |              - value moved here
@@ -120,24 +110,16 @@ LL |         drop(x);
    = note: move occurs because `x` has type `Box<i32>`, which does not implement the `Copy` trait
 
 error[E0382]: use of moved value: `x`
-  --> $DIR/borrowck-multiple-captures.rs:49:19
+  --> $DIR/borrowck-multiple-captures.rs:51:14
    |
-LL |     let x: Box<_> = Box::new(1);
-   |         - move occurs because `x` has type `Box<i32>`, which does not implement the `Copy` trait
-LL |     drop(x);
-   |          - value moved here
-LL |     thread::spawn(move|| {
-   |                   ^^^^^^ value used here after move
-LL |
 LL |         drop(x);
-   |              - use occurs due to use in closure
+   |              - value moved here
+LL |         drop(x);
+   |              ^ value used here after move
    |
-help: consider cloning the value if the performance cost is acceptable
-   |
-LL |     drop(x.clone());
-   |           ++++++++
+   = note: move occurs because `x` has type `Box<i32>`, which does not implement the `Copy` trait
 
-error: aborting due to 8 previous errors
+error: aborting due to 7 previous errors
 
 Some errors have detailed explanations: E0382, E0505.
 For more information about an error, try `rustc --explain E0382`.
