+++ tests/ui/borrowck/issue-103624.stderr
@@ -1,3 +1,22 @@
+error[E0521]: borrowed data escapes outside of method
+  --> $DIR/issue-103624.rs:14:9
+   |
+LL |       async fn foo(&self) {
+   |                    -----
+   |                    |
+   |                    `self` is a reference that is only valid in the method body
+   |                    let's call the lifetime of this reference `'1`
+LL |           let bar = self.b.bar().await;
+LL | /         spawn_blocking(move || {
+LL | |
+LL | |             self.b;
+LL | |
+LL | |         })
+   | |          ^
+   | |          |
+   | |__________`self` escapes the method body here
+   |            argument requires that `'1` must outlive `'static`
+
 error[E0507]: cannot move out of `self.b`, as `self` is a captured variable in an `Fn` closure
   --> $DIR/issue-103624.rs:16:13
    |
@@ -26,25 +45,6 @@ LL |             self.b;
 ...
 LL | struct StructB {}
    | ^^^^^^^^^^^^^^ consider implementing `Clone` for this type
-
-error[E0521]: borrowed data escapes outside of method
-  --> $DIR/issue-103624.rs:14:9
-   |
-LL |       async fn foo(&self) {
-   |                    -----
-   |                    |
-   |                    `self` is a reference that is only valid in the method body
-   |                    let's call the lifetime of this reference `'1`
-LL |           let bar = self.b.bar().await;
-LL | /         spawn_blocking(move || {
-LL | |
-LL | |             self.b;
-LL | |
-LL | |         })
-   | |          ^
-   | |          |
-   | |__________`self` escapes the method body here
-   |            argument requires that `'1` must outlive `'static`
 
 error: aborting due to 2 previous errors
