+++ tests/ui/closures/wrong-closure-arg-suggestion-125325.stderr
@@ -28,15 +28,6 @@ LL |     func(|| x = ())
    |     |    in this closure
    |     expects `Fn` instead of `FnMut`
 
-error[E0594]: cannot assign to `self.counter`, as `Fn` closures cannot mutate their captured variables
-  --> $DIR/wrong-closure-arg-suggestion-125325.rs:46:28
-   |
-LL |         take(|_| to_fn(|_| self.counter += 1));
-   |                  ----- --- ^^^^^^^^^^^^^^^^^ cannot assign
-   |                  |     |
-   |                  |     in this closure
-   |                  expects `Fn` instead of `FnMut`
-
 error: lifetime may not live long enough
   --> $DIR/wrong-closure-arg-suggestion-125325.rs:46:18
    |
@@ -62,14 +53,13 @@ LL | fn take<F, U>(_: F)
    |                  - change this to accept `FnMut` instead of `Fn`
 
 error[E0594]: cannot assign to `self.counter`, as `Fn` closures cannot mutate their captured variables
-  --> $DIR/wrong-closure-arg-suggestion-125325.rs:54:34
+  --> $DIR/wrong-closure-arg-suggestion-125325.rs:46:28
    |
-LL |         P.flat_map(|_| P.map(|_| self.counter += 1));
-   |                          --------^^^^^^^^^^^^^^^^^-
-   |                          |   |   |
-   |                          |   |   cannot assign
-   |                          |   in this closure
-   |                          expects `Fn` instead of `FnMut`
+LL |         take(|_| to_fn(|_| self.counter += 1));
+   |                  ----- --- ^^^^^^^^^^^^^^^^^ cannot assign
+   |                  |     |
+   |                  |     in this closure
+   |                  expects `Fn` instead of `FnMut`
 
 error: lifetime may not live long enough
   --> $DIR/wrong-closure-arg-suggestion-125325.rs:54:24
@@ -96,6 +86,16 @@ LL |         P.flat_map(|_| P.map(|_| self.counter += 1));
    |           |        |         cannot borrow as mutable
    |           |        in this closure
    |           expects `Fn` instead of `FnMut`
+
+error[E0594]: cannot assign to `self.counter`, as `Fn` closures cannot mutate their captured variables
+  --> $DIR/wrong-closure-arg-suggestion-125325.rs:54:34
+   |
+LL |         P.flat_map(|_| P.map(|_| self.counter += 1));
+   |                          --------^^^^^^^^^^^^^^^^^-
+   |                          |   |   |
+   |                          |   |   cannot assign
+   |                          |   in this closure
+   |                          expects `Fn` instead of `FnMut`
 
 error: aborting due to 8 previous errors
