+++ tests/ui/coroutine/auto-trait-regions.stderr
@@ -1,3 +1,15 @@
+error: implementation of `Foo` is not general enough
+  --> $DIR/auto-trait-regions.rs:31:5
+   |
+LL |     let generator = #[coroutine] move || {
+   |                                  ------- this coroutine captures a value whose type is not `Foo`
+...
+LL |     assert_foo(generator);
+   |     ^^^^^^^^^^^^^^^^^^^^^ implementation of `Foo` is not general enough
+   |
+   = note: `&'0 OnlyFooIfStaticRef` must implement `Foo`, for any lifetime `'0`...
+   = note: ...but `Foo` is actually implemented for the type `&'static OnlyFooIfStaticRef`
+
 error[E0626]: borrow may still be in use when coroutine yields
   --> $DIR/auto-trait-regions.rs:45:19
    |
@@ -29,18 +41,6 @@ help: add `static` to mark this coroutine as unmovable
    |
 LL |     let generator = #[coroutine] static move || {
    |                                  ++++++
-
-error: implementation of `Foo` is not general enough
-  --> $DIR/auto-trait-regions.rs:31:5
-   |
-LL |     let generator = #[coroutine] move || {
-   |                                  ------- this coroutine captures a value whose type is not `Foo`
-...
-LL |     assert_foo(generator);
-   |     ^^^^^^^^^^^^^^^^^^^^^ implementation of `Foo` is not general enough
-   |
-   = note: `&'0 OnlyFooIfStaticRef` must implement `Foo`, for any lifetime `'0`...
-   = note: ...but `Foo` is actually implemented for the type `&'static OnlyFooIfStaticRef`
 
 error: implementation of `Foo` is not general enough
   --> $DIR/auto-trait-regions.rs:51:5
