+++ tests/ui/lifetimes/issue-79187-2.stderr
@@ -1,21 +1,3 @@
-error: lifetime may not live long enough
-  --> $DIR/issue-79187-2.rs:11:24
-   |
-LL |     take_foo(|a: &i32| a);
-   |                  -   - ^ returning this value requires that `'1` must outlive `'2`
-   |                  |   |
-   |                  |   return type of closure is &'2 i32
-   |                  let's call the lifetime of this reference `'1`
-
-error: lifetime may not live long enough
-  --> $DIR/issue-79187-2.rs:14:34
-   |
-LL |     take_foo(|a: &i32| -> &i32 { a });
-   |                  -        -      ^ returning this value requires that `'1` must outlive `'2`
-   |                  |        |
-   |                  |        let's call the lifetime of this reference `'2`
-   |                  let's call the lifetime of this reference `'1`
-
 error: implementation of `FnOnce` is not general enough
   --> $DIR/issue-79187-2.rs:8:5
    |
@@ -56,6 +38,15 @@ note: the lifetime requirement is introduced here
 LL | fn take_foo(_: impl Foo) {}
    |                     ^^^
 
+error: lifetime may not live long enough
+  --> $DIR/issue-79187-2.rs:11:24
+   |
+LL |     take_foo(|a: &i32| a);
+   |                  -   - ^ returning this value requires that `'1` must outlive `'2`
+   |                  |   |
+   |                  |   return type of closure is &'2 i32
+   |                  let's call the lifetime of this reference `'1`
+
 error[E0308]: mismatched types
   --> $DIR/issue-79187-2.rs:14:5
    |
@@ -69,6 +60,15 @@ note: the lifetime requirement is introduced here
    |
 LL | fn take_foo(_: impl Foo) {}
    |                     ^^^
+
+error: lifetime may not live long enough
+  --> $DIR/issue-79187-2.rs:14:34
+   |
+LL |     take_foo(|a: &i32| -> &i32 { a });
+   |                  -        -      ^ returning this value requires that `'1` must outlive `'2`
+   |                  |        |
+   |                  |        let's call the lifetime of this reference `'2`
+   |                  let's call the lifetime of this reference `'1`
 
 error: aborting due to 6 previous errors
