+++ tests/ui/macros/return_from_external_macro.stderr
@@ -1,14 +1,3 @@
-error[E0515]: cannot return reference to temporary value
-  --> $DIR/return_from_external_macro.rs:7:13
-   |
-LL |     drop(|| ret_from_ext::foo!());
-   |             ^^^^^^^^^^^^^^^^^^^^
-   |             |
-   |             returns a reference to data owned by the current function
-   |             temporary value created here
-   |
-   = note: this error originates in the macro `ret_from_ext::foo` (in Nightly builds, run with -Z macro-backtrace for more info)
-
 error[E0716]: temporary value dropped while borrowed
   --> $DIR/return_from_external_macro.rs:10:5
    |
@@ -20,6 +9,17 @@ LL |     ret_from_ext::foo!()
 LL |
 LL | }
    | - temporary value is freed at the end of this statement
+   |
+   = note: this error originates in the macro `ret_from_ext::foo` (in Nightly builds, run with -Z macro-backtrace for more info)
+
+error[E0515]: cannot return reference to temporary value
+  --> $DIR/return_from_external_macro.rs:7:13
+   |
+LL |     drop(|| ret_from_ext::foo!());
+   |             ^^^^^^^^^^^^^^^^^^^^
+   |             |
+   |             returns a reference to data owned by the current function
+   |             temporary value created here
    |
    = note: this error originates in the macro `ret_from_ext::foo` (in Nightly builds, run with -Z macro-backtrace for more info)
