+++ tests/ui/pin-ergonomics/pattern-matching-deref-pattern.normal.stderr
@@ -1,23 +1,4 @@
 error[E0507]: cannot move out of a shared reference
-  --> $DIR/pattern-matching-deref-pattern.rs:48:28
-   |
-LL |         let Foo { x, y } = foo.as_mut();
-   |                   -  -     ^^^^^^^^^^^^
-   |                   |  |
-   |                   |  ...and here
-   |                   data moved here
-   |
-   = note: move occurs because these variables have types that don't implement the `Copy` trait
-help: consider borrowing the pattern binding
-   |
-LL |         let Foo { ref x, y } = foo.as_mut();
-   |                   +++
-help: consider borrowing the pattern binding
-   |
-LL |         let Foo { x, ref y } = foo.as_mut();
-   |                      +++
-
-error[E0507]: cannot move out of a shared reference
   --> $DIR/pattern-matching-deref-pattern.rs:38:24
    |
 LL |     let Foo { x, y } = foo.as_mut();
@@ -37,9 +18,9 @@ LL |     let Foo { x, ref y } = foo.as_mut();
    |                  +++
 
 error[E0507]: cannot move out of a shared reference
-  --> $DIR/pattern-matching-deref-pattern.rs:70:28
+  --> $DIR/pattern-matching-deref-pattern.rs:48:28
    |
-LL |         let Foo { x, y } = foo.as_ref();
+LL |         let Foo { x, y } = foo.as_mut();
    |                   -  -     ^^^^^^^^^^^^
    |                   |  |
    |                   |  ...and here
@@ -48,11 +29,11 @@ LL |         let Foo { x, y } = foo.as_ref();
    = note: move occurs because these variables have types that don't implement the `Copy` trait
 help: consider borrowing the pattern binding
    |
-LL |         let Foo { ref x, y } = foo.as_ref();
+LL |         let Foo { ref x, y } = foo.as_mut();
    |                   +++
 help: consider borrowing the pattern binding
    |
-LL |         let Foo { x, ref y } = foo.as_ref();
+LL |         let Foo { x, ref y } = foo.as_mut();
    |                      +++
 
 error[E0507]: cannot move out of a shared reference
@@ -75,23 +56,23 @@ LL |     let Foo { x, ref y } = foo.as_ref();
    |                  +++
 
 error[E0507]: cannot move out of a shared reference
-  --> $DIR/pattern-matching-deref-pattern.rs:92:25
+  --> $DIR/pattern-matching-deref-pattern.rs:70:28
    |
-LL |         let Bar(x, y) = bar.as_mut();
-   |                 -  -    ^^^^^^^^^^^^
-   |                 |  |
-   |                 |  ...and here
-   |                 data moved here
+LL |         let Foo { x, y } = foo.as_ref();
+   |                   -  -     ^^^^^^^^^^^^
+   |                   |  |
+   |                   |  ...and here
+   |                   data moved here
    |
    = note: move occurs because these variables have types that don't implement the `Copy` trait
 help: consider borrowing the pattern binding
    |
-LL |         let Bar(ref x, y) = bar.as_mut();
-   |                 +++
+LL |         let Foo { ref x, y } = foo.as_ref();
+   |                   +++
 help: consider borrowing the pattern binding
    |
-LL |         let Bar(x, ref y) = bar.as_mut();
-   |                    +++
+LL |         let Foo { x, ref y } = foo.as_ref();
+   |                      +++
 
 error[E0507]: cannot move out of a shared reference
   --> $DIR/pattern-matching-deref-pattern.rs:82:21
@@ -113,9 +94,9 @@ LL |     let Bar(x, ref y) = bar.as_mut();
    |                +++
 
 error[E0507]: cannot move out of a shared reference
-  --> $DIR/pattern-matching-deref-pattern.rs:114:25
+  --> $DIR/pattern-matching-deref-pattern.rs:92:25
    |
-LL |         let Bar(x, y) = bar.as_ref();
+LL |         let Bar(x, y) = bar.as_mut();
    |                 -  -    ^^^^^^^^^^^^
    |                 |  |
    |                 |  ...and here
@@ -124,11 +105,11 @@ LL |         let Bar(x, y) = bar.as_ref();
    = note: move occurs because these variables have types that don't implement the `Copy` trait
 help: consider borrowing the pattern binding
    |
-LL |         let Bar(ref x, y) = bar.as_ref();
+LL |         let Bar(ref x, y) = bar.as_mut();
    |                 +++
 help: consider borrowing the pattern binding
    |
-LL |         let Bar(x, ref y) = bar.as_ref();
+LL |         let Bar(x, ref y) = bar.as_mut();
    |                    +++
 
 error[E0507]: cannot move out of a shared reference
@@ -149,6 +130,25 @@ help: consider borrowing the pattern binding
    |
 LL |     let Bar(x, ref y) = bar.as_ref();
    |                +++
+
+error[E0507]: cannot move out of a shared reference
+  --> $DIR/pattern-matching-deref-pattern.rs:114:25
+   |
+LL |         let Bar(x, y) = bar.as_ref();
+   |                 -  -    ^^^^^^^^^^^^
+   |                 |  |
+   |                 |  ...and here
+   |                 data moved here
+   |
+   = note: move occurs because these variables have types that don't implement the `Copy` trait
+help: consider borrowing the pattern binding
+   |
+LL |         let Bar(ref x, y) = bar.as_ref();
+   |                 +++
+help: consider borrowing the pattern binding
+   |
+LL |         let Bar(x, ref y) = bar.as_ref();
+   |                    +++
 
 error: aborting due to 8 previous errors
