+++ tests/ui/suggestions/option-content-move3.stderr
@@ -1,29 +1,4 @@
 error[E0507]: cannot move out of `var`, a captured variable in an `FnMut` closure
-  --> $DIR/option-content-move3.rs:13:21
-   |
-LL |     let var = NotCopyable;
-   |         --- captured outer variable
-...
-LL |         move || {
-   |         ------- captured by this `FnMut` closure
-LL |             let x = var;
-   |                     ^^^ move occurs because `var` has type `NotCopyable`, which does not implement the `Copy` trait
-   |
-   = help: `Fn` and `FnMut` closures require captured values to be able to be consumed multiple times, but `FnOnce` closures may consume them only once
-note: if `NotCopyable` implemented `Clone`, you could clone the value
-  --> $DIR/option-content-move3.rs:2:1
-   |
-LL | struct NotCopyable;
-   | ^^^^^^^^^^^^^^^^^^ consider implementing `Clone` for this type
-...
-LL |             let x = var;
-   |                     --- you could clone this value
-help: consider borrowing here
-   |
-LL |             let x = &var;
-   |                     +
-
-error[E0507]: cannot move out of `var`, a captured variable in an `FnMut` closure
   --> $DIR/option-content-move3.rs:12:9
    |
 LL |     let var = NotCopyable;
@@ -53,17 +28,25 @@ LL |             let x = var;
    |                     --- you could clone this value
 
 error[E0507]: cannot move out of `var`, a captured variable in an `FnMut` closure
-  --> $DIR/option-content-move3.rs:24:21
+  --> $DIR/option-content-move3.rs:13:21
    |
-LL |     let var = NotCopyableButCloneable;
+LL |     let var = NotCopyable;
    |         --- captured outer variable
 ...
 LL |         move || {
    |         ------- captured by this `FnMut` closure
 LL |             let x = var;
-   |                     ^^^ move occurs because `var` has type `NotCopyableButCloneable`, which does not implement the `Copy` trait
+   |                     ^^^ move occurs because `var` has type `NotCopyable`, which does not implement the `Copy` trait
    |
    = help: `Fn` and `FnMut` closures require captured values to be able to be consumed multiple times, but `FnOnce` closures may consume them only once
+note: if `NotCopyable` implemented `Clone`, you could clone the value
+  --> $DIR/option-content-move3.rs:2:1
+   |
+LL | struct NotCopyable;
+   | ^^^^^^^^^^^^^^^^^^ consider implementing `Clone` for this type
+...
+LL |             let x = var;
+   |                     --- you could clone this value
 help: consider borrowing here
    |
 LL |             let x = &var;
@@ -99,6 +82,23 @@ LL |             println!("{x:?}");
 LL ~         }
 LL +         }
    |
+
+error[E0507]: cannot move out of `var`, a captured variable in an `FnMut` closure
+  --> $DIR/option-content-move3.rs:24:21
+   |
+LL |     let var = NotCopyableButCloneable;
+   |         --- captured outer variable
+...
+LL |         move || {
+   |         ------- captured by this `FnMut` closure
+LL |             let x = var;
+   |                     ^^^ move occurs because `var` has type `NotCopyableButCloneable`, which does not implement the `Copy` trait
+   |
+   = help: `Fn` and `FnMut` closures require captured values to be able to be consumed multiple times, but `FnOnce` closures may consume them only once
+help: consider borrowing here
+   |
+LL |             let x = &var;
+   |                     +
 
 error: aborting due to 4 previous errors
